FEAT: boolq benchmark done.

+++ eval/models/load.py
@@ -1,11 +1,16 @@
 from configs import config
 from transformers import AutoModelForCausalLM, AutoTokenizer
+import torch
 
 def load_model(model_path: str = None):
     try:
         if not model_path:
             model_path = config['model_path']
-        model = AutoModelForCausalLM.from_pretrained(model_path, device_map="auto")
+        model = AutoModelForCausalLM.from_pretrained(
+                    model_path, 
+                    device_map="auto",
+                    torch_dtype=torch.float16
+        )
         return model
     except ValueError as value_error:
         print(f"ValueError: {value_error} - Check if the model path is correct and contains valid model files.")
